requirements.txt
streamlit
mysql-connector-python
pandas
plotly
numpy
python-dotenv
